Add files via upload
The create_prime_numbers_mpy.py is my porting of my calculate_prime_numbers.py file into MicroPython to run on an RP2040 microcontroller. I had to tweak working with the libraries, and the code. I added a date-time stamp for each entry written to the file, and I invoked the onboard LED to show when the program was running. I limited the output to prime numbers up to one thousand as the storage space on the RP2040 is limited.